requirements.txt
flask
pandas
openpyxl
pdfplumber
werkzeug
openai
python-dotenv
flask-cors
gunicorn
